+++ src/RecoveryServices/RecoveryServices.Backup.Models/Properties/Resources.Designer.cs
@@ -1926,6 +1926,76 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Properties {
             }
         }
         /// <summary>
+        ///   Looks up a localized string similar to Both -IsSystemAssignedIdentity and -UserAssignedIdentityArmUrl were specified. Specify only one identity for identity-based Azure Files access.
+        /// </summary>
+        public static string AFSIdentityBothSpecified
+        {
+            get
+            {
+                return ResourceManager.GetString("AFSIdentityBothSpecified", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to -AccessType &apos;IdentityBased&apos; requires an identity.
+        /// </summary>
+        public static string AFSIdentityBasedRequiresIdentity
+        {
+            get
+            {
+                return ResourceManager.GetString("AFSIdentityBasedRequiresIdentity", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to An identity was specified without -AccessType. Specify -AccessType 'IdentityBased' when using -IsSystemAssignedIdentity or -UserAssignedIdentityArmUrl.
+        /// </summary>
+        public static string AFSIdentityRequiresAccessType
+        {
+            get
+            {
+                return ResourceManager.GetString("AFSIdentityRequiresAccessType", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to -AccessType &apos;KeyBased&apos; cannot be combined with an identity.
+        /// </summary>
+        public static string AFSKeyBasedWithIdentity
+        {
+            get
+            {
+                return ResourceManager.GetString("AFSKeyBasedWithIdentity", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to The storage account &apos;{0}&apos; is already registered with a different access type / identity. Re-registering will change how backup accesses the storage account. Do you want to continue?
+        /// </summary>
+        public static string AFSReregisterIdentityChangeWarning
+        {
+            get
+            {
+                return ResourceManager.GetString("AFSReregisterIdentityChangeWarning", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to Re-register storage account with new access type / identity.
+        /// </summary>
+        public static string AFSReregisterIdentityChangeCaption
+        {
+            get
+            {
+                return ResourceManager.GetString("AFSReregisterIdentityChangeCaption", resourceCulture);
+            }
+        }
+        /// <summary>
+        ///   Looks up a localized string similar to Re-registration of the storage account was canceled by the user.
+        /// </summary>
+        public static string AFSReregisterCanceled
+        {
+            get
+            {
+                return ResourceManager.GetString("AFSReregisterCanceled", resourceCulture);
+            }
+        }
+        /// <summary>
         ///   Looks up a localized string similar to Please ensure that the Recovery point is Zone pinned or remove TargetZone.
         /// </summary>
         public static string RecoveryPointZonePinnedException